+++ functions.json
@@ -17,7 +17,7 @@
         }
       }
     },
-    "function_type": "Mutation",
+    "functionType": "Mutation",
     "identifier": "messages.js:send",
     "returns": {
       "type": "null"
@@ -30,7 +30,7 @@
     "args": {
       "type": "any"
     },
-    "function_type": "Query",
+    "functionType": "Query",
     "identifier": "messages.js:list",
     "returns": {
       "type": "any"
